ocr_service/front/requirements.txt
streamlit
request
 #pip install -r requirements.txt